Updated PsiCashLib framework to XCFramework bundle type
