--- PsiCashLib.framework/build-git-commit.txt
@@ -1,1 +1,0 @@
-v2.0.0-beta.4-0-g3538c00
