+++ PsiCashLib.xcframework/ios-arm64_i386_x86_64-simulator/PsiCashLib.framework/Headers/PsiCashLib.h
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2020, Psiphon Inc.
+ * All rights reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+*/
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for PsiCashLib.
+FOUNDATION_EXPORT double PsiCashLibVersionNumber;
+
+//! Project version string for PsiCashLib.
+FOUNDATION_EXPORT const unsigned char PsiCashLibVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <PsiCashLib/PublicHeader.h>
+
+#import <PsiCashLib/PsiCashLibWrapper.h>
