+++ PsiCashLib.xcframework/ios-arm64_i386_x86_64-simulator/PsiCashLib.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module PsiCashLib {
+  umbrella header "PsiCashLib.h"
+
+  export *
+  module * { export * }
+}
