+++ PsiCashLib.xcframework/ios-arm64_i386_x86_64-simulator/PsiCashLib.framework/build-git-commit.txt
@@ -1,0 +1,1 @@
+v2.0.0-beta.6-0-g453d7f7
